mingw: allow for longer paths in `parse_interpreter()`
As reported in https://github.com/newren/git-filter-repo/pull/225, it
looks like 99 bytes is not really sufficient to represent e.g. the full
path to Python when installed via Windows Store (and this path is used
in the hasb bang line when installing scripts via `pip`).

Let's increase it to what is probably the maximum sensible path size:
248 (the same as for `CreateDirectory()`).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>